requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator>=2.2.0
openai==1.3.0
anthropic==0.7.0
numpy>=1.26.0
google-generativeai>=0.8.0
requests==2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

